Detect and re-enable queue if scaling in fails (fixes #37)
For both SGE and Slurm, when attempting to scale in a cluster which has
the minimum number of compute nodes for its auto-scaling group and there
are no pending or running jobs, the autoscaler will detect that it needs
to scale in, attempt to do so, but then fail as the last node cannot be
shut down. We then do not detect this and never re-enable the node's
queue, leaving it stuck unusable until manual action is taken.

The solution is to return the exit code of AWS scaling in so that we can
tell this has happened and reenable the queue. Note this needs to be
done for both SGE and Slurm until #36 has been done.