Remove unneeded flight/session-* handler hooks
These aren't needed, and have never been needed, they were just a proof
of concept.

FlightDeck currently has nothing to do with Clusterware sessions, and
there are no plans to add this at the moment. If we do later want to add
similar handling to what these hooks provided, it would likely be for
the AAM appliance and the hooks would live in the cluster-appliances
handler; I have made a note in case we want to adapt these and add them
back in there.