Ensure config written if only access master
Previously the appliance config would only be written if we were the
storage master node. If we're the storage master node we tend to be the
access master node as well, but this is not guaranteed, so we should
ensure the config is written if either one (or both) of these is the
case.