Store output under hostname
My previous assumption that the queue was only processed on the master node
was wrong.  The periodic cronjob only runs on the master node, but it
broadcasts to all nodes in the cluster, so they all run the job queue handler.

This caused some race conditions:

 1. The logs and status files for all nodes were being stored in the same
    location.  So only the last logs were being kept.
 2. The job may not be reliably ran on all nodes.  The first node to reject or
    process the job deletes it from s3.  This could potentially be before
    other nodes have downloaded it from s3.
 3. The job itself may have side effects which do not support concurrent runs
    of the job.  This could result in unexpected behaviour.

This commit addresses the first of these issues by including the hostname
in the key for the logs and status files.