Run jobs only on master node
Previously I had assumed that the queue was only processed on the master node.
This was not the case.  The periodic cronjob only runs on the master node, but
it broadcasts to all nodes in the cluster, so they were all running the job
queue handler.

This caused some race conditions:

 1. The logs and status files for all nodes were being stored in the same
    location.  So only the last logs were being kept.
 2. The job may not be reliably ran on all nodes.  The first node to reject or
    process the job deletes it from s3.  This could potentially be before
    other nodes have downloaded it from s3.
 3. The job itself may have side effects which do not support concurrent runs
    of the job.  This could result in unexpected behaviour.

A previous commit addressed the first of these issues by including the
hostname in the key for the logs and status files.

This commit addresses points 2 and 3.  The job queue handler is now only ran
for the master node.  If there are multiple master nodes in a cluster, issues
2 and 3 will, or could, still present.