Update genders file with compute group membership information.
We add an `autoscaling-member-join` handler to `clusterable` so that, when a
node in a compute group joins the cluster, that information is recorded in the
cluster genders file.

Removing the entry when the node leaves is handled by the existing
`member-leave` handler (which removes _all_ lines in the genders file for a
departing node).