Fix job queues not running
At some point recently the variable `cw_INSTANCE_role`, used to
determine if this node should process job queues, has become undefined
when this script is executed (though presumably was defined in the past
when this executed as this used to work); this therefore caused job
queues to never be processed by nodes as the condition would never be
true.

Explicitly loading the config file defining this variable appears to fix
this. Presumably the file was being loaded via some dependency somehow
before, and this has since changed. I had a look but couldn't see an
obvious cause of this, possibly this is related to the recent
Customizers refactoring or additions?