Allow running job queue jobs on any node
We now upload job queue jobs with the node the job is intended for
included in the folder name, in order to allow running jobs on any node
by querying the appropriate folder. In order to process jobs on other
nodes we also need to remove this check that the current node is the
master node.