Assume a node has a single slot if it's otherwise unspecified.
If a node joins the cluster before it's had its node resource tags
written, then it won't have a value for `slots` when this script is
called.

Slurm doesn't like that, so we'll just assume it's got a single slot for
now. When the tags are written, the slave's serf daemon will be
restarted and we'll get a new `member-join` with accurate information.